# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth API for Connect.",
        "title": "Connect Auth API",
        "contact": {
            "email": "mrexmelle@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/credentials": {
            "post": {
                "description": "Post a new credential",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credentials"
                ],
                "parameters": [
                    {
                        "description": "Credential Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/credential.PostRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/credential.ResponseDto"
                        }
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            }
        },
        "/credentials/{employee_id}": {
            "delete": {
                "description": "Delete a credential",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credentials"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "employee_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/credential.ResponseDto"
                        }
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            }
        },
        "/credentials/{employee_id}/password": {
            "delete": {
                "description": "Reset a credential's password",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credentials"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "employee_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/credential.ResponseDto"
                        }
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            },
            "patch": {
                "description": "Patch a credential's password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credentials"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "employee_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Password Patch Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/credential.PatchPasswordRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/credential.ResponseDto"
                        }
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            }
        },
        "/profiles/{ehid}": {
            "get": {
                "description": "Get a profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee Hash ID",
                        "name": "ehid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/profile.ResponseDto"
                        }
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            },
            "delete": {
                "description": "Delete a profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee Hash ID",
                        "name": "ehid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/profile.DeleteResponseDto"
                        }
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            },
            "patch": {
                "description": "Patch a profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee Hash ID",
                        "name": "ehid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Profile Patch Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profile.PatchRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/profile.PatchResponseDto"
                        }
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            }
        }
    },
    "definitions": {
        "credential.PatchPasswordRequestDto": {
            "type": "object",
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                }
            }
        },
        "credential.PostRequestDto": {
            "type": "object",
            "properties": {
                "employee_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "credential.ResponseDto": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "profile.DeleteResponseDto": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "profile.Entity": {
            "type": "object",
            "properties": {
                "dob": {
                    "type": "string"
                },
                "ehid": {
                    "type": "string"
                },
                "email_address": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "profile.PatchRequestDto": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "profile.PatchResponseDto": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "profile.ResponseDto": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/profile.Entity"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}